Move firewall to awsmodel
