+++ pkg/model/awsmodel/api_loadbalancer.go
@@ -308,7 +308,7 @@ func (b *APILoadBalancerBuilder) Build(c *fi.ModelBuilderContext) error {
 			Egress:        fi.Bool(true),
 			SecurityGroup: lbSG,
 		}
-		b.AddDirectionalGroupRule(c, t)
+		AddDirectionalGroupRule(c, t)
 	}
 
 	// Allow traffic into the ELB from KubernetesAPIAccess CIDRs
@@ -323,7 +323,7 @@ func (b *APILoadBalancerBuilder) Build(c *fi.ModelBuilderContext) error {
 				SecurityGroup: lbSG,
 				ToPort:        fi.Int64(443),
 			}
-			b.AddDirectionalGroupRule(c, t)
+			AddDirectionalGroupRule(c, t)
 
 			// Allow ICMP traffic required for PMTU discovery
 			c.AddTask(&awstasks.SecurityGroupRule{
@@ -356,7 +356,7 @@ func (b *APILoadBalancerBuilder) Build(c *fi.ModelBuilderContext) error {
 					SecurityGroup: masterGroup.Task,
 					ToPort:        fi.Int64(443),
 				}
-				b.AddDirectionalGroupRule(c, t)
+				AddDirectionalGroupRule(c, t)
 
 				// Allow ICMP traffic required for PMTU discovery
 				c.AddTask(&awstasks.SecurityGroupRule{
